Fix? saving subforms
Something was still barfing in strange corner cases, so the whole save
routine was scrapped. Now, it does the following:

1. Saves the form, `commit=False`
2. Saves the formsets, `commit=False`, and collects the returned
   instances. This handles deletion for us.
3. Saves the related forms, `commit=False`, and collects the returned
   instance (if any). This handles deletion for us.
4. Make some helper methods that set the instance on the related
   instances, and save them.
5. If `commit`, save the instance and the related instances using the
   helpers.
6. Otherwise, expose the helpers so they can be called later, after the
   instance is saved.
7. Return the instance.

This has the benefit of being only one path of code. All code uses
the helpers to save the instances. The only point of difference is who
calls the relevant methods - the form itself, or the calling code.

Additionally, the foreign key is set in a different place, and matches
the [example code][1] given in the Django docs for this kind of
situation.

[1]: https://docs.djangoproject.com/en/dev/topics/forms/modelforms/#saving-objects-in-the-formset